Added in support for getting location by zip code
also tidied up a bit so that current location grabs from api response not menu. However, this make it so that it doesn't include the state. would that really be an issue? I'd also just like you (ken) to check over what I've wrote on the copy before we put it into the main file.